+++ backend/app/Config/PublicConfig.php
@@ -69,6 +69,7 @@ class PublicConfig extends ConfigFactory
             ConfigInterface::APP_BACKGROUND_TYPE_LOCK => 'false',
 
             ConfigInterface::BRANDING_SHOW_POWERED_BY => 'true',
+            ConfigInterface::BRANDING_SHOW_VERSION => 'true',
             ConfigInterface::CHATBOT_DISPLAY_NAME => '',
             ConfigInterface::CHATBOT_AVATAR_URL => '',
             ConfigInterface::SIDEBAR_NAVIGATION_CONFIG => '',